rename: interface폴더 이름 변경, constant폴더 생성
interface라는 이름이 types관련 변수 파일이라는 것을 명확히 알 수 없다고 생각하여 types로 변경후 카테고리 상 lib폴더 아래에 두었습니다. constants.ts 파일도 여러 constant 파일이 생성 될 수 있다고 판단이 들어 constant폴더를 생성후 lib폴더 아래에 두었습니다
