+++ src/components/Dialog/DeleteDialog.tsx
@@ -1,7 +1,7 @@
 import { useModalHandleContext } from 'contexts/ModalContext';
 import { useHabitDispatchContext } from 'contexts/HabitContext';
 import Dialog from './Dialog';
-import type { Habit } from 'interface/main';
+import type { Habit } from 'lib/types/main';
 
 const DeleteDialog = ({ habit }: { habit: Habit }) => {
   const { closeModal } = useModalHandleContext();
